[refactor]: remove usage of deprecated ListView
React Native V 0.60 no longer has a ListView component included. We need to either migrate to use FlatList or alias usage of ListView to deprecated-react-native-listview. Seeing as we have already forked the Dropdown component to make it support Web, it makes sense to just refactor it to remove the usage of ListView.